EC-54 Test and implement `check_heatnetwork_present`
This commit implements the check_heatnetwork_present method as part of
the notional FHS wrapper. It also creates a unit test covering this
method using the test input file. Note, that this test did not exist in
the python codebase, however it has helped to provide confidence in the
Rust port/implementation.