EC-54 add skylight and roof to notional test data
This commits adds two building elements (a skylight and a roof) to the
test data used by the notional fhs wrapper tests. This is so we can
bring the edit_glazing_for_glazing_limit under test which modifies
input data on the these types of building elements only.

Note: The equivalent of this test does not exist in the HEM Python.
We temporarily created it so we can duplicate it in Rust which this
commit begins to implement.

Co-Authored-By: Russ Penska <russ-madetech@users.noreply.github.com>
Co-Authored-By: Rachel Howell <98879818+rachowell@users.noreply.github.com>
Co-Authored-By: Eleanor Mollett <18617190+emollett@users.noreply.github.com>