EC-469 Port SmartHotWaterTank demand_hot_water test
This commit adds the first half of the last SmartHotWaterTank tests:
test_demand_hot_water_for_smart_hot_water_tank

Note that it currently fails and has therefore been ignored.

The reason it fails is that the thermostat_layer (correctly!) is None.
In Python this can be None as the method this is eventually passed into
is on SmartHotWaterTank and does not rely on the thermostat_layer
(_determine_heat_source_switch_on). While in Rust we currently
incorrectly call the determine_heat_source_switch_on on the StorageTank.