EC-1039 Refactor BoilerServiceWaterCombi tests
This commit moves and renames fixtures to more closely match the
Python test set up. It creates a submodule inside the test module
matching the test class used in Python. It also renames a test to
match the Python test name so it is easier to identify which tests are
which.